+++ packages/balm-core/src/bundler/rollup/plugins.ts
@@ -1,48 +1,37 @@
-import { nodeResolve } from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
 import { terser } from 'rollup-plugin-terser';
 import { InputOptions, OutputOptions, RollupPlugin } from '@balm-core/index';
 
 function getInputPlugins(customInputOptions: InputOptions): RollupPlugin[] {
-  const {
-    inputOptions,
-    nodeResolveOptions,
-    commonjsOptions
-  } = BalmJS.config.scripts;
+  const { inputOptions } = BalmJS.config.scripts;
 
-  let inputPlugins = [
-    nodeResolve(
-      Object.assign(
-        {
-          customResolveOptions: {
-            moduleDirectory: 'node_modules'
-          },
-          preferBuiltins: false
-        },
-        nodeResolveOptions
-      )
-    ),
-    commonjs(commonjsOptions)
-  ];
+  const inputPlugins: RollupPlugin[] = [];
 
   if (inputOptions.plugins) {
-    inputPlugins = inputPlugins.concat(inputOptions.plugins);
+    inputPlugins.push(...inputOptions.plugins);
   }
 
   if (customInputOptions.plugins) {
-    inputPlugins = inputPlugins.concat(customInputOptions.plugins);
+    inputPlugins.push(...customInputOptions.plugins);
   }
 
   return inputPlugins;
 }
 
 function getOutputPlugins(customOutputOptions: OutputOptions): RollupPlugin[] {
-  let outputPlugins = BalmJS.config.env.isProd
-    ? [terser(BalmJS.config.scripts.minifyOptions)]
-    : [];
+  const { outputOptions } = BalmJS.config.scripts;
+
+  const outputPlugins: RollupPlugin[] = [];
+
+  if (outputOptions.plugins) {
+    outputPlugins.push(...outputOptions.plugins);
+  }
+
+  if (BalmJS.config.env.isProd) {
+    outputPlugins.push(terser(BalmJS.config.scripts.minifyOptions));
+  }
 
   if (customOutputOptions.plugins) {
-    outputPlugins = outputPlugins.concat(customOutputOptions.plugins);
+    outputPlugins.push(...customOutputOptions.plugins);
   }
 
   return outputPlugins;